+++ TestProject/ViewModels/ProductViewModel.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Title { get; set; }
         public int Quantity { get; set; }
-        public decimal Price { get; set; }
-        public decimal TotalPriceWithVAT { get; set; }
+        public double Price { get; set; }
+        public double TotalPriceWithVAT { get; set; }
     }
 }
